requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.26.2
